Create a ratings panel
