+++ src/scripts/clipperUI/mainController.tsx
@@ -31,6 +31,7 @@ import {DialogButton, DialogPanel} from "./panels/dialogPanel";
 import {ErrorDialogPanel} from "./panels/errorDialogPanel";
 import {LoadingPanel} from "./panels/loadingPanel";
 import {OptionsPanel} from "./panels/optionsPanel";
+import {RatingsPanel} from "./panels/ratingsPanel";
 import {RegionSelectingPanel} from "./panels/regionSelectingPanel";
 import {SignInPanel} from "./panels/signInPanel";
 import {SuccessPanel} from "./panels/successPanel";
@@ -282,7 +283,13 @@ export class MainControllerClass extends ComponentBase<MainControllerState, Main
 				return <ErrorDialogPanel message={OneNoteApiUtils.getLocalizedErrorMessage(apiResponseCode) }
 					buttons={buttons} />;
 			case PanelType.ClippingSuccess:
-				return <SuccessPanel clipperState={this.props.clipperState} />;
+				let panels: any[] = [<SuccessPanel clipperState={this.props.clipperState} />];
+
+				if (this.props.clipperState.shouldShowRatingsPrompt) {
+					panels.push(<RatingsPanel clipperState={this.props.clipperState} />);
+				}
+
+				return panels;
 			default:
 				return undefined;
 		}
