+++ src/scripts/clipperUI/panels/successPanel.tsx
@@ -9,26 +9,11 @@ import {ClipMode} from "../clipMode";
 import {ClipperStateProp} from "../clipperState";
 import {ComponentBase} from "../componentBase";
 import {Clipper} from "../frontEndGlobals";
-import {RatingsHelper, RatingsPromptStage} from "../ratingsHelper";
 import {Status} from "../status";
 
 import {SpriteAnimation} from "../components/spriteAnimation";
 
-import {DialogButton, DialogPanel} from "./dialogPanel";
-
-interface SuccessPanelState {
-	currentRatingsPromptStage?: RatingsPromptStage;
-	userSelectedRatingsPromptStage?: RatingsPromptStage;
-}
-
-export class SuccessPanelClass extends ComponentBase<SuccessPanelState, ClipperStateProp> {
-	getInitialState(): SuccessPanelState {
-		return {
-			currentRatingsPromptStage: RatingsPromptStage.INIT,
-			userSelectedRatingsPromptStage: RatingsPromptStage.INIT
-		};
-	}
-
+class SuccessPanelClass extends ComponentBase<{ }, ClipperStateProp> {
 	public onLaunchOneNoteButton() {
 		Clipper.logger.logUserFunnel(Log.Funnel.Label.ViewInWac);
 		let data = this.props.clipperState.oneNoteApiResult.data as OneNoteApi.Page;
@@ -40,47 +25,24 @@ export class SuccessPanelClass extends ComponentBase<SuccessPanelState, ClipperS
 		}
 	}
 
-	private showRatingsPrompt(): any[] {
-		if (this.props.clipperState.shouldShowRatingsPrompt) {
-			let message: string = RatingsHelper.getMessage(this.state.userSelectedRatingsPromptStage);
-			let buttons: DialogButton[] = RatingsHelper.getDialogButtons(this);
-
-			if (!Utils.isNullOrUndefined(message)) {
-				let animationStrategy = RatingsHelper.getAnimationStategy(this);
-
-				return (
-					<DialogPanel
-						message={message}
-						buttons={buttons}
-						buttonFontFamily={Localization.FontFamily.Regular}
-						divId={Constants.Ids.ratingsPromptContainer}
-						animationStrategy={animationStrategy} />
-				);
-			}
-		}
-	}
-
 	render() {
 		return (
 			<div id={Constants.Ids.clipperSuccessContainer}>
-				<div>
-					<div className="messageLabelContainer successPagePadding">
-						<SpriteAnimation spriteUrl={Utils.getImageResourceUrl("checkmark.png")} imageHeight={28} totalFrameCount={30} loop={false}/>
-						<span className="actionLabelFont messageLabel"
-							style={Localization.getFontFamilyAsStyle(Localization.FontFamily.Light)}>
-							{Localization.getLocalizedString("WebClipper.Label.ClipSuccessful")}
+				<div className="messageLabelContainer successPagePadding">
+					<SpriteAnimation spriteUrl={Utils.getImageResourceUrl("checkmark.png")} imageHeight={28} totalFrameCount={30} loop={false}/>
+					<span className="actionLabelFont messageLabel"
+						style={Localization.getFontFamilyAsStyle(Localization.FontFamily.Light)}>
+						{Localization.getLocalizedString("WebClipper.Label.ClipSuccessful")}
+					</span>
+				</div>
+				<a id={Constants.Ids.launchOneNoteButton} {...this.enableInvoke(this.onLaunchOneNoteButton, 70) }>
+					<div className="wideButtonContainer">
+						<span className="wideButtonFont wideActionButton"
+							style={Localization.getFontFamilyAsStyle(Localization.FontFamily.Regular)}>
+							{Localization.getLocalizedString("WebClipper.Action.ViewInOneNote")}
 						</span>
 					</div>
-					<a id={Constants.Ids.launchOneNoteButton} {...this.enableInvoke(this.onLaunchOneNoteButton, 70) }>
-						<div className="wideButtonContainer">
-							<span className="wideButtonFont wideActionButton"
-								style={Localization.getFontFamilyAsStyle(Localization.FontFamily.Regular)}>
-								{Localization.getLocalizedString("WebClipper.Action.ViewInOneNote")}
-							</span>
-						</div>
-					</a>
-				</div>
-				{this.showRatingsPrompt()}
+				</a>
 			</div>
 		);
 	}
